Remove un-used v1 stuff
